+++ src/lib.rs
@@ -281,7 +281,7 @@ use num_traits::{One, Zero};
 use std::collections::{BinaryHeap, HashMap};
 use std::hash::Hash;
 use std::iter;
-use std::ops::{AddAssign, SubAssign};
+use std::ops::SubAssign;
 #[cfg(test)]
 mod unit_tests;
 
@@ -328,46 +328,9 @@ where
     }
 }
 
-impl<T, N> Counter<T, N>
-where
-    T: Hash + Eq,
-    N: Zero,
-{
-    /// Create a new, empty `Counter`
-    pub fn new() -> Counter<T, N> {
-        Counter {
-            map: HashMap::new(),
-            zero: N::zero(),
-        }
-    }
-}
 
-impl<T, N> Counter<T, N>
-where
-    T: Hash + Eq,
-    N: AddAssign + Zero + One,
-{
-    /// Create a new `Counter` initialized with the given iterable.
-    pub fn init<I>(iterable: I) -> Counter<T, N>
-    where
-        I: IntoIterator<Item = T>,
-    {
-        let mut counter = Counter::new();
-        counter.update(iterable);
-        counter
-    }
 
-    /// Add the counts of the elements from the given iterable to this counter.
-    pub fn update<I>(&mut self, iterable: I)
-    where
-        I: IntoIterator<Item = T>,
-    {
-        for item in iterable {
-            let entry = self.map.entry(item).or_insert_with(N::zero);
-            *entry += N::one();
-        }
-    }
-}
+
 
 impl<T, N> Counter<T, N>
 where
